move execution engine to a separate file
