+++ arbnode/transaction_streamer.go
@@ -21,8 +21,6 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/math"
-	"github.com/ethereum/go-ethereum/core"
-	"github.com/ethereum/go-ethereum/core/state"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethdb"
 	"github.com/ethereum/go-ethereum/log"
@@ -30,8 +28,6 @@ import (
 	"github.com/ethereum/go-ethereum/rlp"
 
 	"github.com/offchainlabs/nitro/arbos"
-	"github.com/offchainlabs/nitro/arbos/arbosState"
-	"github.com/offchainlabs/nitro/arbos/l1pricing"
 	"github.com/offchainlabs/nitro/arbstate"
 	"github.com/offchainlabs/nitro/arbutil"
 	"github.com/offchainlabs/nitro/broadcaster"
@@ -39,25 +35,6 @@ import (
 	"github.com/offchainlabs/nitro/util/sharedmetrics"
 	"github.com/offchainlabs/nitro/util/stopwaiter"
 )
-
-type ExecutionEngine struct {
-	stopwaiter.StopWaiter
-
-	bc        *core.BlockChain
-	validator *staker.BlockValidator
-	streamer  *TransactionStreamer
-
-	resequenceChan    chan []*arbstate.MessageWithMetadata
-	createBlocksMutex sync.Mutex
-
-	newBlockNotifier chan struct{}
-	latestBlockMutex sync.Mutex
-	latestBlock      *types.Block
-
-	nextScheduledVersionCheck time.Time // protected by the createBlocksMutex
-
-	reorgSequencing func() *arbos.SequencingHooks
-}
 
 // TransactionStreamer produces blocks from a node's L1 messages, storing the results in the blockchain and recording their positions
 // The streamer is notified when there's new batches to process
@@ -103,14 +80,6 @@ func TransactionStreamerConfigAddOptions(prefix string, f *flag.FlagSet) {
 	f.Int64(prefix+".max-reorg-resequence-depth", DefaultTransactionStreamerConfig.MaxReorgResequenceDepth, "maximum number of messages to attempt to resequence on reorg (0 = never resequence, -1 = always resequence)")
 }
 
-func NewExecutionEngine(bc *core.BlockChain) (*ExecutionEngine, error) {
-	return &ExecutionEngine{
-		bc:               bc,
-		resequenceChan:   make(chan []*arbstate.MessageWithMetadata),
-		newBlockNotifier: make(chan struct{}, 1),
-	}, nil
-}
-
 func NewTransactionStreamer(
 	db ethdb.Database,
 	chainConfig *params.ChainConfig,
@@ -145,16 +114,6 @@ func uint64ToKey(x uint64) []byte {
 	return data
 }
 
-func (s *ExecutionEngine) setBlockValidator(validator *staker.BlockValidator) {
-	if s.Started() {
-		panic("trying to set block validator after start")
-	}
-	if s.validator != nil {
-		panic("trying to set block validator when already set")
-	}
-	s.validator = validator
-}
-
 // TODO: this is needed only for block validator
 func (s *TransactionStreamer) SetBlockValidator(validator *staker.BlockValidator) {
 	s.exec.setBlockValidator(validator)
@@ -178,16 +137,6 @@ func (s *TransactionStreamer) SetInboxReader(inboxReader *InboxReader) {
 		panic("trying to set inbox reader when already set")
 	}
 	s.inboxReader = inboxReader
-}
-
-func (s *ExecutionEngine) SetReorgSequencingPolicy(reorgSequencing func() *arbos.SequencingHooks) {
-	if s.Started() {
-		panic("trying to set reorg sequencing policy after start")
-	}
-	if s.reorgSequencing != nil {
-		panic("trying to set reorg sequencing policy when already set")
-	}
-	s.reorgSequencing = reorgSequencing
 }
 
 func (s *TransactionStreamer) cleanupInconsistentState() error {
@@ -332,46 +281,6 @@ func (s *TransactionStreamer) reorg(batch ethdb.Batch, count arbutil.MessageInde
 	}
 
 	return setMessageCount(batch, count)
-}
-
-func (s *ExecutionEngine) Reorg(count arbutil.MessageIndex, newMessages []arbstate.MessageWithMetadata, oldMessages []*arbstate.MessageWithMetadata) error {
-	s.createBlocksMutex.Lock()
-	successful := false
-	defer func() {
-		if !successful {
-			s.createBlocksMutex.Unlock()
-		}
-	}()
-	blockNum, err := s.MessageCountToBlockNumber(count)
-	if err != nil {
-		return err
-	}
-	// We can safely cast blockNum to a uint64 as we checked count == 0 above
-	targetBlock := s.bc.GetBlockByNumber(uint64(blockNum))
-	if targetBlock == nil {
-		log.Warn("reorg target block not found", "block", blockNum)
-		return nil
-	}
-	if s.validator != nil {
-		err = s.validator.ReorgToBlock(targetBlock.NumberU64(), targetBlock.Hash())
-		if err != nil {
-			return err
-		}
-	}
-
-	err = s.bc.ReorgToOldBlock(targetBlock)
-	if err != nil {
-		return err
-	}
-	for i := range newMessages {
-		err := s.digestMessageWithBlockMutex(count+arbutil.MessageIndex(i), &newMessages[i])
-		if err != nil {
-			return err
-		}
-	}
-	s.resequenceChan <- oldMessages
-	successful = true
-	return nil
 }
 
 func setMessageCount(batch ethdb.KeyValueWriter, count arbutil.MessageIndex) error {
@@ -690,24 +599,6 @@ func (s *TransactionStreamer) logReorg(pos arbutil.MessageIndex, dbMsg *arbstate
 
 }
 
-func (s *ExecutionEngine) HeadMessageNumber() (arbutil.MessageIndex, error) {
-	msgCount, err := s.BlockNumberToMessageCount(s.bc.CurrentBlock().Header().Number.Uint64())
-	if err != nil {
-		return 0, err
-	}
-	return msgCount - 1, err
-}
-
-func (s *ExecutionEngine) HeadMessageNumberSync(t *testing.T) (arbutil.MessageIndex, error) {
-	s.createBlocksMutex.Lock()
-	defer s.createBlocksMutex.Unlock()
-	return s.HeadMessageNumber()
-}
-
-func (s *ExecutionEngine) NextDelayedMessageNumber() (uint64, error) {
-	return s.bc.CurrentHeader().Nonce.Uint64(), nil
-}
-
 func (s *TransactionStreamer) addMessagesAndEndBatchImpl(messageStartPos arbutil.MessageIndex, messagesAreConfirmed bool, messages []arbstate.MessageWithMetadata, batch ethdb.Batch) error {
 	var confirmedReorg bool
 	var oldMsg *arbstate.MessageWithMetadata
@@ -857,153 +748,6 @@ func messageFromTxes(header *arbos.L1IncomingMessageHeader, txes types.Transacti
 	}, nil
 }
 
-// The caller must hold the createBlocksMutex
-func (s *ExecutionEngine) resequenceReorgedMessages(messages []*arbstate.MessageWithMetadata) {
-	if s.reorgSequencing == nil {
-		return
-	}
-
-	log.Info("Trying to resequence messages", "number", len(messages))
-	lastBlockHeader := s.bc.CurrentBlock().Header()
-	if lastBlockHeader == nil {
-		log.Error("block header not found during resequence")
-		return
-	}
-
-	nextDelayedSeqNum := lastBlockHeader.Nonce.Uint64()
-
-	for _, msg := range messages {
-		// Check if the message is non-nil just to be safe
-		if msg == nil || msg.Message == nil || msg.Message.Header == nil {
-			continue
-		}
-		header := msg.Message.Header
-		if header.RequestId != nil {
-			delayedSeqNum := header.RequestId.Big().Uint64()
-			if delayedSeqNum != nextDelayedSeqNum {
-				log.Info("not resequencing delayed message due to unexpected index", "expected", nextDelayedSeqNum, "found", delayedSeqNum)
-				continue
-			}
-			err := s.sequenceDelayedMessageWithBlockMutex(msg.Message, delayedSeqNum)
-			if err != nil {
-				log.Error("failed to re-sequence old delayed message removed by reorg", "err", err)
-			}
-			nextDelayedSeqNum += 1
-			continue
-		}
-		if header.Kind != arbos.L1MessageType_L2Message || header.Poster != l1pricing.BatchPosterAddress {
-			// This shouldn't exist?
-			log.Warn("skipping non-standard sequencer message found from reorg", "header", header)
-			continue
-		}
-		// We don't need a batch fetcher as this is an L2 message
-		txes, err := msg.Message.ParseL2Transactions(s.bc.Config().ChainID, nil)
-		if err != nil {
-			log.Warn("failed to parse sequencer message found from reorg", "err", err)
-			continue
-		}
-		_, err = s.sequenceTransactionsWithBlockMutex(msg.Message.Header, txes, s.reorgSequencing())
-		if err != nil {
-			log.Error("failed to re-sequence old user message removed by reorg", "err", err)
-			return
-		}
-	}
-}
-
-func (s *ExecutionEngine) SequenceTransactions(header *arbos.L1IncomingMessageHeader, txes types.Transactions, hooks *arbos.SequencingHooks) (*types.Block, error) {
-	for {
-		hooks.TxErrors = nil
-		s.createBlocksMutex.Lock()
-		block, err := s.sequenceTransactionsWithBlockMutex(header, txes, hooks)
-		s.createBlocksMutex.Unlock()
-		if !errors.Is(err, ErrSequencerInsertLockTaken) {
-			return block, err
-		}
-		<-time.After(time.Millisecond * 100)
-	}
-}
-
-func (s *ExecutionEngine) sequenceTransactionsWithBlockMutex(header *arbos.L1IncomingMessageHeader, txes types.Transactions, hooks *arbos.SequencingHooks) (*types.Block, error) {
-	lastBlockHeader := s.bc.CurrentBlock().Header()
-	if lastBlockHeader == nil {
-		return nil, errors.New("current block header not found")
-	}
-
-	statedb, err := s.bc.StateAt(lastBlockHeader.Root)
-	if err != nil {
-		return nil, err
-	}
-
-	delayedMessagesRead := lastBlockHeader.Nonce.Uint64()
-
-	startTime := time.Now()
-	block, receipts, err := arbos.ProduceBlockAdvanced(
-		header,
-		txes,
-		delayedMessagesRead,
-		lastBlockHeader,
-		statedb,
-		s.bc,
-		s.bc.Config(),
-		hooks,
-	)
-	if err != nil {
-		return nil, err
-	}
-	blockCalcTime := time.Since(startTime)
-	if len(hooks.TxErrors) != len(txes) {
-		return nil, fmt.Errorf("unexpected number of error results: %v vs number of txes %v", len(hooks.TxErrors), len(txes))
-	}
-
-	if len(receipts) == 0 {
-		return nil, nil
-	}
-
-	allTxsErrored := true
-	for _, err := range hooks.TxErrors {
-		if err == nil {
-			allTxsErrored = false
-			break
-		}
-	}
-	if allTxsErrored {
-		return nil, nil
-	}
-
-	msg, err := messageFromTxes(header, txes, hooks.TxErrors)
-	if err != nil {
-		return nil, err
-	}
-
-	msgWithMeta := arbstate.MessageWithMetadata{
-		Message:             msg,
-		DelayedMessagesRead: delayedMessagesRead,
-	}
-
-	pos, err := s.BlockNumberToMessageCount(lastBlockHeader.Number.Uint64())
-	if err != nil {
-		return nil, err
-	}
-
-	err = s.streamer.WriteMessageFromSequencer(pos, msgWithMeta)
-	if err != nil {
-		return nil, err
-	}
-
-	// Only write the block after we've written the messages, so if the node dies in the middle of this,
-	// it will naturally recover on startup by regenerating the missing block.
-	err = s.appendBlock(block, statedb, receipts, blockCalcTime)
-	if err != nil {
-		return nil, err
-	}
-
-	if s.validator != nil {
-		s.validator.NewBlock(block, lastBlockHeader, msgWithMeta)
-	}
-
-	return block, nil
-}
-
 var ErrSequencerInsertLockTaken = errors.New("insert lock taken")
 
 func (s *TransactionStreamer) WriteMessageFromSequencer(pos arbutil.MessageIndex, msgWithMeta arbstate.MessageWithMetadata) error {
@@ -1045,80 +789,8 @@ func (s *TransactionStreamer) WriteMessageFromSequencer(pos arbutil.MessageIndex
 	return nil
 }
 
-func (s *ExecutionEngine) SequenceDelayedMessage(message *arbos.L1IncomingMessage, delayedSeqNum uint64) error {
-	for {
-		s.createBlocksMutex.Lock()
-		err := s.sequenceDelayedMessageWithBlockMutex(message, delayedSeqNum)
-		s.createBlocksMutex.Unlock()
-		if !errors.Is(err, ErrSequencerInsertLockTaken) {
-			return err
-		}
-		<-time.After(time.Millisecond * 100)
-	}
-}
-
-func (s *ExecutionEngine) sequenceDelayedMessageWithBlockMutex(message *arbos.L1IncomingMessage, delayedSeqNum uint64) error {
-	currentHeader := s.bc.CurrentBlock().Header()
-
-	expectedDelayed := currentHeader.Nonce.Uint64()
-
-	pos, err := s.BlockNumberToMessageCount(currentHeader.Number.Uint64())
-	if err != nil {
-		return err
-	}
-
-	if expectedDelayed != delayedSeqNum {
-		return fmt.Errorf("wrong delayed message sequenced got %d expected %d", delayedSeqNum, expectedDelayed)
-	}
-
-	messageWithMeta := arbstate.MessageWithMetadata{
-		Message:             message,
-		DelayedMessagesRead: delayedSeqNum + 1,
-	}
-
-	err = s.streamer.WriteMessageFromSequencer(pos, messageWithMeta)
-	if err != nil {
-		return err
-	}
-
-	startTime := time.Now()
-	block, statedb, receipts, err := s.createBlockFromNextMessage(&messageWithMeta)
-	if err != nil {
-		return err
-	}
-
-	err = s.appendBlock(block, statedb, receipts, time.Since(startTime))
-	if err != nil {
-		return err
-	}
-
-	log.Info("ExecutionEngine: Added DelayedMessages", "pos", pos, "delayed", delayedSeqNum, "block-header", block.Header())
-
-	return nil
-}
-
 func (s *TransactionStreamer) GetGenesisBlockNumber() (uint64, error) {
 	return s.chainConfig.ArbitrumChainParams.GenesisBlockNum, nil
-}
-
-func (s *ExecutionEngine) GetGenesisBlockNumber() (uint64, error) {
-	return s.bc.Config().ArbitrumChainParams.GenesisBlockNum, nil
-}
-
-func (s *ExecutionEngine) BlockNumberToMessageCount(blockNum uint64) (arbutil.MessageIndex, error) {
-	genesis, err := s.GetGenesisBlockNumber()
-	if err != nil {
-		return 0, err
-	}
-	return arbutil.BlockNumberToMessageCount(blockNum, genesis), nil
-}
-
-func (s *ExecutionEngine) MessageCountToBlockNumber(messageNum arbutil.MessageIndex) (int64, error) {
-	genesis, err := s.GetGenesisBlockNumber()
-	if err != nil {
-		return 0, err
-	}
-	return arbutil.MessageCountToBlockNumber(messageNum, genesis), nil
 }
 
 // PauseReorgs until a matching call to ResumeReorgs (may be called concurrently)
@@ -1205,133 +877,6 @@ func (s *TransactionStreamer) feedNextMsg(ctx context.Context, exec *ExecutionEn
 	return pos+1 < msgCount
 }
 
-// must hold createBlockMutex
-func (s *ExecutionEngine) createBlockFromNextMessage(msg *arbstate.MessageWithMetadata) (*types.Block, *state.StateDB, types.Receipts, error) {
-	currentHeader := s.bc.CurrentBlock().Header()
-	if currentHeader == nil {
-		return nil, nil, nil, errors.New("failed to get current header")
-	}
-
-	statedb, err := s.bc.StateAt(currentHeader.Root)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	statedb.StartPrefetcher("TransactionStreamer")
-	defer statedb.StopPrefetcher()
-
-	batchFetcher := func(batchNum uint64) ([]byte, error) {
-		return s.streamer.inboxReader.GetSequencerMessageBytes(context.TODO(), batchNum)
-	}
-
-	block, receipts, err := arbos.ProduceBlock(
-		msg.Message,
-		msg.DelayedMessagesRead,
-		currentHeader,
-		statedb,
-		s.bc,
-		s.bc.Config(),
-		batchFetcher,
-	)
-
-	return block, statedb, receipts, err
-}
-
-// must hold createBlockMutex
-func (s *ExecutionEngine) appendBlock(block *types.Block, statedb *state.StateDB, receipts types.Receipts, duration time.Duration) error {
-	var logs []*types.Log
-	for _, receipt := range receipts {
-		logs = append(logs, receipt.Logs...)
-	}
-	status, err := s.bc.WriteBlockAndSetHeadWithTime(block, receipts, logs, statedb, true, duration)
-	if err != nil {
-		return err
-	}
-	if status == core.SideStatTy {
-		return errors.New("geth rejected block as non-canonical")
-	}
-	return nil
-}
-
-func (s *ExecutionEngine) DigestMessage(num arbutil.MessageIndex, msg *arbstate.MessageWithMetadata) error {
-	if !s.createBlocksMutex.TryLock() {
-		return errors.New("mutex held")
-	}
-	defer s.createBlocksMutex.Unlock()
-	return s.digestMessageWithBlockMutex(num, msg)
-}
-
-func (s *ExecutionEngine) digestMessageWithBlockMutex(num arbutil.MessageIndex, msg *arbstate.MessageWithMetadata) error {
-	currentHeader := s.bc.CurrentHeader()
-	expNum, err := s.BlockNumberToMessageCount(currentHeader.Number.Uint64())
-	if err != nil {
-		return err
-	}
-	if expNum != num {
-		return fmt.Errorf("wrong message number in digest got %d expected %d", num, expNum)
-	}
-
-	startTime := time.Now()
-	block, statedb, receipts, err := s.createBlockFromNextMessage(msg)
-	if err != nil {
-		return err
-	}
-
-	err = s.appendBlock(block, statedb, receipts, time.Since(startTime))
-	if err != nil {
-		return err
-	}
-
-	if s.validator != nil {
-		s.validator.NewBlock(block, currentHeader, *msg)
-	}
-
-	if time.Now().After(s.nextScheduledVersionCheck) {
-		s.nextScheduledVersionCheck = time.Now().Add(time.Minute)
-		arbState, err := arbosState.OpenSystemArbosState(statedb, nil, true)
-		if err != nil {
-			return err
-		}
-		version, timestampInt, err := arbState.GetScheduledUpgrade()
-		if err != nil {
-			return err
-		}
-		var timeUntilUpgrade time.Duration
-		var timestamp time.Time
-		if timestampInt == 0 {
-			// This upgrade will take effect in the next block
-			timestamp = time.Now()
-		} else {
-			// This upgrade is scheduled for the future
-			timestamp = time.Unix(int64(timestampInt), 0)
-			timeUntilUpgrade = time.Until(timestamp)
-		}
-		maxSupportedVersion := params.ArbitrumDevTestChainConfig().ArbitrumChainParams.InitialArbOSVersion
-		logLevel := log.Warn
-		if timeUntilUpgrade < time.Hour*24 {
-			logLevel = log.Error
-		}
-		if version > maxSupportedVersion {
-			logLevel(
-				"you need to update your node to the latest version before this scheduled ArbOS upgrade",
-				"timeUntilUpgrade", timeUntilUpgrade,
-				"upgradeScheduledFor", timestamp,
-				"maxSupportedArbosVersion", maxSupportedVersion,
-				"pendingArbosUpgradeVersion", version,
-			)
-		}
-	}
-
-	sharedmetrics.UpdateSequenceNumberInBlockGauge(num)
-	s.latestBlockMutex.Lock()
-	s.latestBlock = block
-	s.latestBlockMutex.Unlock()
-	select {
-	case s.newBlockNotifier <- struct{}{}:
-	default:
-	}
-	return nil
-}
-
 func (s *TransactionStreamer) Start(ctxIn context.Context) {
 	s.StopWaiter.Start(ctxIn, s)
 	s.LaunchThread(func(ctx context.Context) {
@@ -1348,47 +893,6 @@ func (s *TransactionStreamer) Start(ctxIn context.Context) {
 			case <-timer.C:
 			}
 
-		}
-	})
-}
-
-func (s *ExecutionEngine) Start(ctx_in context.Context) {
-	s.StopWaiter.Start(ctx_in, s)
-	s.LaunchThread(func(ctx context.Context) {
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case resequence := <-s.resequenceChan:
-				s.resequenceReorgedMessages(resequence)
-				s.createBlocksMutex.Unlock()
-			}
-		}
-	})
-	s.LaunchThread(func(ctx context.Context) {
-		var lastBlock *types.Block
-		for {
-			select {
-			case <-s.newBlockNotifier:
-			case <-ctx.Done():
-				return
-			}
-			s.latestBlockMutex.Lock()
-			block := s.latestBlock
-			s.latestBlockMutex.Unlock()
-			if block != lastBlock && block != nil {
-				log.Info(
-					"created block",
-					"l2Block", block.Number(),
-					"l2BlockHash", block.Hash(),
-				)
-				lastBlock = block
-				select {
-				case <-time.After(time.Second):
-				case <-ctx.Done():
-					return
-				}
-			}
 		}
 	})
 }
